ccb extractor. first version
