+++ AzureDataLake/DecisionServiceExtractor.Test/DecisionServiceExtractor.Test.csproj
@@ -65,6 +65,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CcbJsonParserTest.cs" />
     <Compile Include="JsonParserTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
