+++ AzureDataLake/DecisionServiceExtractor/FieldExpression.cs
@@ -1,0 +1,11 @@
+﻿namespace DecisionServiceExtractor
+{
+    class FieldExpression
+    {
+        public string FieldName;
+
+        public string JsonPath;
+
+        public int Idx;
+    }
+}
