+++ AzureDataLake/Scripts/ProcessCCB.usql
@@ -1,0 +1,70 @@
+﻿USE DATABASE [DecisionServiceDBTest];
+
+REFERENCE ASSEMBLY DecisionServiceDBTest.[DecisionServiceExtractor];
+
+DECLARE @_appId string = "xbetdealshomespotlight";    //your appId
+DECLARE @_account string = "decisionengine";    // your account name
+
+DECLARE @_year string = "2019";  //your date: year
+DECLARE @_month string = "10";     //your date: month
+DECLARE @_day string = "14";       //your date: day
+
+DECLARE @OutputFolder string = "/" + @_appId + "_daily/";      //your output folder
+DECLARE @OutputSlimLogs bool = false;     //true if you want to have slime version of logs (all metadata without features)
+     
+DECLARE @InputFile string = "wasb://"+ @_appId +"@" + @_account + ".blob.core.windows.net/{model}/data/{year}/{month}/{day}_{filename}.json";
+
+DECLARE @OutputFileInteractions string = @OutputFolder + "interactions-" + @_year + "-" + @_month + "-" + @_day + ".csv";
+DECLARE @OutputFileDangling string = @OutputFolder + "dangling-" + @_year + "-" + @_month + "-" + @_day + ".csv";
+DECLARE @OutputFileStatistics string = @OutputFolder + "statistics-" + @_year + "-" + @_month + "-" + @_day + ".csv";
+DECLARE @OutputFileInvalid string = @OutputFolder + "invalid-" + @_year + "-" + @_month + "-" + @_day + ".csv";
+
+@Events =  
+    EXTRACT SessionId string,
+            SlotIdx int,
+            EventId string,
+            model string,
+            ParseError string,
+            day string, //virtual column
+            month string, //virtual column
+            year string, //virtual column
+            filename string //virtual column
+    FROM @InputFile
+    USING new DecisionServiceExtractor.CcbExtractor();
+
+@Events = SELECT * FROM @Events WHERE year == @_year AND month == @_month AND day == @_day;
+
+@Valid = SELECT * FROM @Events WHERE string.IsNullOrEmpty(ParseError);
+
+@Invalid = SELECT ParseError FROM @Events WHERE !string.IsNullOrEmpty(ParseError);
+
+
+@Statistics =
+    SELECT COUNT() AS SessionCount
+    FROM @Valid;
+
+OUTPUT @Statistics
+TO @OutputFileStatistics
+USING Outputters.Csv(outputHeader:true);
+
+OUTPUT @Invalid
+TO @OutputFileInvalid
+USING Outputters.Csv(outputHeader:true);
+
+IF @OutputSlimLogs == true THEN
+    @Dangling =
+        SELECT EventId,
+            EnqueuedTimeUtc
+        FROM @Events
+        WHERE IsDangling == true;
+
+    OUTPUT @Interactions
+    TO @OutputFileInteractions
+    USING Outputters.Csv(outputHeader:true);
+
+    OUTPUT @Dangling
+    TO @OutputFileDangling
+    USING Outputters.Csv(outputHeader:true);
+END;
+
+
