+++ AzureDataLake/Scripts/ProcessCCB.usql.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.Analytics.Interfaces;
+using Microsoft.Analytics.Types.Sql;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+
+namespace Scripts
+{
+    
+}
